==> Problem 1/Joy/01. Data Cleaning.ipynb <==
import pandas as pd #importing pandas library as pd, like creating an instance of pandas
import matplotlib.pyplot as plt
%matplotlib inline

df_boxes = pd.read_csv('../Common Datasets/boxes.csv') #importing box details datasets
df_problem1 = pd.read_csv('Datasets/problem 1.csv') #importing predict datasets
df_purchase = pd.read_csv('../Common Datasets/purchase.csv') #importing history datasets

# <font color="red">First checking the boxes datasets, is it clean or not</font>

df_boxes.head() #checking collumn and head of datasets

df_boxes.info() #checking some information

=> no null value found and 2 collumn are in object type

df_boxes.describe() #checking min, max, mean, quartile

df_boxes['BOX_ID'].nunique()

=> so there are 1-290 box in total, no outlier here. full clean

# <font color="red">Checking Problem one dataset</font>

df_problem1.head()

df_problem1.info()

=> so total 58689 MAGIC_KEY, we need to predict for them

# <font color="red">Checking Purchase history</font>

df_purchase.head()

df_purchase.info()

# <font color="red">Checking Null Value</font>

df_purchase.isnull().sum()

=> we have some null value, we need to fix these\
=> as we have enough large datasets, we can drop these\
=> in problem 3, we have a magic key that have null value here, so we need to fill this

len(df_purchase)

box_count_mean = df_purchase['BOX_COUNT'].mean()
box_count_mean = int(box_count_mean)
box_count_mean

df_purchase['BOX_COUNT'] = df_purchase['BOX_COUNT'].fillna(box_count_mean)

BOX_ID_Mode = df_purchase['BOX_ID'].mode()[0]
BOX_ID_Mode

df_purchase['BOX_ID'] = df_purchase['BOX_ID'].fillna(BOX_ID_Mode)

df_purchase.isnull().sum()

# <font color="red">Checking duplicate</font>

duplicates =  df_purchase[df_purchase.duplicated(keep='first')] #it will return the duplicated row from the datasets
len(duplicates)

duplicates

df_purchase.loc[(df_purchase.MAGIC_KEY == '2C3ECE53EE6')] #to show the information of male only

df_purchase = df_purchase.drop_duplicates() #this method will return the dataset with no duplicate row
len(df_purchase)

# <font color="red">Checking outlier</font>

df_purchase.describe()

=> we have some outlier, need to fix these

for column in df_purchase.columns:
    unique_values_count = df_purchase[column].value_counts()
    print(f"Column: {column}")
    print(unique_values_count)
    print()

=> we can see, in BOX_COUNT, there is a value -1, its not possible in real life, its an outlier\
=> 19 is not outlier as other close value still exist. there is no big gap

unique_values_count = df_purchase['BOX_ID'].value_counts()
print(f"Column: {'BOX_ID'}")
print(unique_values_count)
print()

import numpy as np
from scipy import stats

z_scores = np.abs(stats.zscore(df_purchase['BOX_ID']))
threshold = 3
outliers = np.where(z_scores > threshold)[0]
print("Outliers detected using Z-score method:", outliers)

outliers

len(outliers)

df_purchase.iloc[outliers] #same as df.head(), inner third bracet using as a numpy array

=> both BOX_ID and BOX_COUNT value are inappropriate, we can drop these row

df_purchase = df_purchase.drop(df_purchase.index[outliers])
df_purchase

df_purchase.describe()

=> now there is no outlier, MAX for BOX_ID and MIN for BOX_COUNT Fixed

# <font color="red">Now we need to merge the dataset for model</font>

merged_df = pd.merge(df_purchase, df_boxes, on='BOX_ID')

merged_df

merged_df.info()

merged_df.describe()

merged_df['PURCHASE_DATE'] = pd.to_datetime(merged_df['PURCHASE_DATE'], format="%d/%m/%Y")

merged_df.info()

merged_df.describe()

merged_df.to_csv('merged_df.csv', index=False)
