==> Problem 1/Joy/02. Train Test Split.ipynb <==
import pandas as pd

df = pd.read_csv('../Common Datasets/merged_df.csv')

df

df.info()

df['PURCHASE_DATE'] = pd.to_datetime(df['PURCHASE_DATE'])

train_df = df[df['PURCHASE_DATE'] < '2019-02-01']

# Filter the data for February 2019
test_df = df[(df['PURCHASE_DATE'] >= '2019-02-01') & (df['PURCHASE_DATE'] < '2019-03-01')]

# Save the filtered datasets to separate files or data structures
train_df.to_csv("train_df.csv", index=False)  # Save data before February
test_df.to_csv("test_df.csv", index=False)  # Save data for February
