==> Problem 1/Joy/03. TestDataClean.ipynb <==
import pandas as pd

test_df = pd.read_csv('Datasets/test_df.csv')
train_df = pd.read_csv('Datasets/train_df.csv')

len(train_df)

len(test_df)

train_df['MAGIC_KEY'].nunique()

test_df['MAGIC_KEY'].nunique()

new_train_df = train_df.groupby('MAGIC_KEY')['PURCHASE'].apply(lambda x: 1 if (x == 'Y').any() else 0).reset_index()

new_test_df = test_df.groupby('MAGIC_KEY')['PURCHASE'].apply(lambda x: 1 if (x == 'Y').any() else 0).reset_index()

len(new_train_df)

len(new_test_df)

merged_df = pd.merge(new_test_df, new_train_df, on='MAGIC_KEY', how='inner')

len(merged_df)

merged_df

new_test_df.loc[(new_test_df.MAGIC_KEY == '2E6F53ECC13')]

merged_df = merged_df.drop(columns=['PURCHASE_y'])

merged_df

merged_df = merged_df.rename(columns={'PURCHASE_x': 'PURCHASE'})

merged_df

merged_df.to_csv("test.csv", index=False)  # Save test data
