==> Problem 1/Joy/05. New Datasets from merged.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline
from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('../Common Datasets/merged_df.csv')

df

df['MAGIC_KEY'].nunique()

df.info()

df['PURCHASE_DATE'] = pd.to_datetime(df['PURCHASE_DATE'])

# <font color="red">We need to create a new datasets with customer behaviour</font>

# Group by 'MAGIC_KEY' and count the number of purchases for each customer
df_customer = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].count().reset_index()
df_customer.columns = ['MAGIC_KEY', 'PURCHASE_COUNT']

# Display the total number of purchases made by each customer
df_customer

# Calculate the total amount spent for each purchase
df['TOTAL_AMOUNT'] = df['BOX_COUNT'] * df['UNIT_PRICE']

# Group by 'MAGIC_KEY' and sum the total amount spent by each customer
total_amount_spent = df.groupby('MAGIC_KEY')['TOTAL_AMOUNT'].sum().reset_index()
total_amount_spent.columns = ['MAGIC_KEY', 'TOTAL_AMOUNT_SPENT']

# Display the total amount spent by each customer
total_amount_spent

df_customer = pd.merge(df_customer, total_amount_spent, on='MAGIC_KEY')
df_customer

last_purchase_date = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].max().reset_index()
last_purchase_date.columns = ['MAGIC_KEY', 'LAST_PURCHASE_DATE']

# Display the DataFrame with the last purchase date for each customer
last_purchase_date

last_purchase_date.info()

last_purchase_date['DAY_SINCE_LAST_PURCHASE'] = pd.to_datetime('2019-03-01') - last_purchase_date['LAST_PURCHASE_DATE']
last_purchase_date

df_customer = pd.merge(df_customer, last_purchase_date, on='MAGIC_KEY')
df_customer

df_sorted = df.sort_values(['MAGIC_KEY', 'PURCHASE_DATE'])

# Calculate the time difference between consecutive purchases for each customer
df_sorted['TIME_DIFF'] = df_sorted.groupby('MAGIC_KEY')['PURCHASE_DATE'].diff()

# Group by 'MAGIC_KEY' and calculate the average time difference between consecutive purchases
average_diff = df_sorted.groupby('MAGIC_KEY')['TIME_DIFF'].mean().reset_index()
average_diff.columns = ['MAGIC_KEY', 'AVERAGE_DIFF']

# Fill NaN values (customers with only one purchase) with 0
average_diff['AVERAGE_DIFF'] = average_diff['AVERAGE_DIFF'].fillna(0)
average_diff['AVERAGE_DIFF'] = pd.to_timedelta(average_diff['AVERAGE_DIFF'])
average_diff

df_customer = pd.merge(df_customer, average_diff, on='MAGIC_KEY')
df_customer

df_customer.info()

df_customer.describe()

threshold = pd.Timedelta(days=15)
filtered_df = df_customer.loc[(df_customer['PURCHASE_COUNT'] > 1) & (df_customer['AVERAGE_DIFF'] < threshold)]

filtered_df

df_customer.info()

# Filter the DataFrame to include only dates in the first 15 days of October 2018
october_first_15_df = df[(df['PURCHASE_DATE'].dt.year == 2018) & 
                          (df['PURCHASE_DATE'].dt.month == 10) &
                          (df['PURCHASE_DATE'].dt.day <= 15)]

november_first_15_df = df[(df['PURCHASE_DATE'].dt.year == 2018) & 
                          (df['PURCHASE_DATE'].dt.month == 11) &
                          (df['PURCHASE_DATE'].dt.day <= 15)]

december_first_15_df = df[(df['PURCHASE_DATE'].dt.year == 2018) & 
                          (df['PURCHASE_DATE'].dt.month == 12) &
                          (df['PURCHASE_DATE'].dt.day <= 15)]

january_first_15_df = df[(df['PURCHASE_DATE'].dt.year == 2019) & 
                          (df['PURCHASE_DATE'].dt.month == 1) &
                          (df['PURCHASE_DATE'].dt.day <= 15)]

february_first_15_df = df[(df['PURCHASE_DATE'].dt.year == 2019) & 
                          (df['PURCHASE_DATE'].dt.month == 2) &
                          (df['PURCHASE_DATE'].dt.day <= 15)]

# Define a function to check if there is a purchase in the specified date range
def check_purchase(df, month, year):
    column_name = f"{month}_{year}_FIRST_15"
    return df.groupby('MAGIC_KEY')['PURCHASE_DATE'].count().astype(bool).astype(int).rename(column_name)

# Apply the function to each sub dataset
october_2018_first_15 = check_purchase(october_first_15_df, '10', '2018')
november_2018_first_15 = check_purchase(november_first_15_df, '11', '2018')
december_2018_first_15 = check_purchase(december_first_15_df, '12', '2018')
january_2019_first_15 = check_purchase(january_first_15_df, '1', '2019')
february_2019_first_15 = check_purchase(february_first_15_df, '2', '2019')

df_customer = pd.merge(df_customer, october_2018_first_15, on='MAGIC_KEY', how='left')
df_customer = pd.merge(df_customer, november_2018_first_15, on='MAGIC_KEY', how='left')
df_customer = pd.merge(df_customer, december_2018_first_15, on='MAGIC_KEY', how='left')
df_customer = pd.merge(df_customer, january_2019_first_15, on='MAGIC_KEY', how='left')
df_customer = pd.merge(df_customer, february_2019_first_15, on='MAGIC_KEY', how='left')

df_customer['10_2018_FIRST_15'] = df_customer['10_2018_FIRST_15'].fillna(0)
df_customer['11_2018_FIRST_15'] = df_customer['11_2018_FIRST_15'].fillna(0)
df_customer['12_2018_FIRST_15'] = df_customer['12_2018_FIRST_15'].fillna(0)
df_customer['1_2019_FIRST_15'] = df_customer['1_2019_FIRST_15'].fillna(0)
df_customer['2_2019_FIRST_15'] = df_customer['2_2019_FIRST_15'].fillna(0)

df_customer

df_customer.to_csv("df_customer_all.csv", index=False)  # Save test data
