==> Problem 1/Joy/08. New Features from train.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline
from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('Datasets/train_df.csv')

df

df.info()

df['PURCHASE_DATE'] = pd.to_datetime(df['PURCHASE_DATE'])

# <font color="red">We need to create a new datasets with customer behaviour</font>

df_customer = pd.read_csv('Datasets/df_customer_train.csv')

df_customer.info()

df_customer['LAST_PURCHASE_DATE'] = pd.to_datetime(df_customer['LAST_PURCHASE_DATE'])
df_customer['DAY_SINCE_LAST_PURCHASE'] = pd.to_timedelta(df_customer['DAY_SINCE_LAST_PURCHASE'])
df_customer['AVERAGE_DIFF'] = pd.to_timedelta(df_customer['AVERAGE_DIFF'])

df_customer.info()

df_customer.describe()

october_2018_last_15 = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].apply(
    lambda x: 1 if ((x.dt.month == 10) & (x.dt.day > 15) & (x.dt.year == 2018)).any() else 0
).reset_index(name='10_2018_LAST_15')

# Display the result DataFrame
october_2018_last_15

df_customer = pd.merge(df_customer, october_2018_last_15, on='MAGIC_KEY')
df_customer

november_2018_last_15 = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].apply(
    lambda x: 1 if ((x.dt.month == 11) & (x.dt.day > 15) & (x.dt.year == 2018)).any() else 0
).reset_index(name='11_2018_LAST_15')

# Display the result DataFrame
november_2018_last_15

df_customer = pd.merge(df_customer, november_2018_last_15, on='MAGIC_KEY')
df_customer

december_2018_last_15 = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].apply(
    lambda x: 1 if ((x.dt.month == 12) & (x.dt.day > 15) & (x.dt.year == 2018)).any() else 0
).reset_index(name='12_2018_LAST_15')

# Display the result DataFrame
december_2018_last_15

df_customer = pd.merge(df_customer, december_2018_last_15, on='MAGIC_KEY')
df_customer

january_2018_last_15 = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].apply(
    lambda x: 1 if ((x.dt.month == 1) & (x.dt.day > 15) & (x.dt.year == 2019)).any() else 0
).reset_index(name='1_2019_LAST_15')

# Display the result DataFrame
january_2018_last_15

df_customer = pd.merge(df_customer, january_2018_last_15, on='MAGIC_KEY')
df_customer

df_customer.to_csv("df_customer_train2.csv", index=False)  # Save test data
