==> Problem 1/Joy/09. New Features from merged.ipynb <==
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline
from sklearn.preprocessing import LabelEncoder

df = pd.read_csv('Datasets/merged_df.csv')

df

df.info()

df['PURCHASE_DATE'] = pd.to_datetime(df['PURCHASE_DATE'])

# <font color="red">We need to create a new datasets with customer behaviour</font>

# Group by 'MAGIC_KEY' and count the number of purchases for each customer
df_customer = df.groupby('MAGIC_KEY')['PURCHASE'].count().reset_index()
df_customer.columns = ['MAGIC_KEY', 'PURCHASE_COUNT']

# Display the total number of purchases made by each customer
df_customer

# Calculate the total amount spent for each purchase
df['TOTAL_AMOUNT'] = df['BOX_COUNT'] * df['UNIT_PRICE']

# Group by 'MAGIC_KEY' and sum the total amount spent by each customer
total_amount_spent = df.groupby('MAGIC_KEY')['TOTAL_AMOUNT'].sum().reset_index()
total_amount_spent.columns = ['MAGIC_KEY', 'TOTAL_AMOUNT_SPENT']

# Display the total amount spent by each customer
total_amount_spent

df_customer = pd.merge(df_customer, total_amount_spent, on='MAGIC_KEY')
df_customer

last_purchase_date = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].max().reset_index()
last_purchase_date.columns = ['MAGIC_KEY', 'LAST_PURCHASE_DATE']

# Display the DataFrame with the last purchase date for each customer
last_purchase_date

last_purchase_date.info()

last_purchase_date['DAY_SINCE_LAST_PURCHASE'] = pd.to_datetime('2019-03-01') - last_purchase_date['LAST_PURCHASE_DATE']
last_purchase_date

df_customer = pd.merge(df_customer, last_purchase_date, on='MAGIC_KEY')
df_customer

df_sorted = df.sort_values(['MAGIC_KEY', 'PURCHASE_DATE'])

# Calculate the time difference between consecutive purchases for each customer
df_sorted['TIME_DIFF'] = df_sorted.groupby('MAGIC_KEY')['PURCHASE_DATE'].diff()

# Group by 'MAGIC_KEY' and calculate the average time difference between consecutive purchases
average_diff = df_sorted.groupby('MAGIC_KEY')['TIME_DIFF'].mean().reset_index()
average_diff.columns = ['MAGIC_KEY', 'AVERAGE_DIFF']

# Fill NaN values (customers with only one purchase) with 0
average_diff['AVERAGE_DIFF'] = average_diff['AVERAGE_DIFF'].fillna(0)
average_diff['AVERAGE_DIFF'] = pd.to_timedelta(average_diff['AVERAGE_DIFF'])
average_diff

df_customer = pd.merge(df_customer, average_diff, on='MAGIC_KEY')
df_customer

df_customer.info()

df_customer.describe()

threshold = pd.Timedelta(days=15)
filtered_df = df_customer.loc[(df_customer['PURCHASE_COUNT'] > 1) & (df_customer['AVERAGE_DIFF'] < threshold)]

filtered_df

#new_purchase_df = df.groupby('MAGIC_KEY')['PURCHASE'].apply(lambda x: 1 if (x == 'Y').any() else 0).reset_index()
#new_purchase_df

#df_customer = pd.merge(df_customer, new_purchase_df, on='MAGIC_KEY')
#df_customer

#le = LabelEncoder() #creating an instance of LabelEncoder

#df_encoded = df_customer.copy()
#df_encoded['MAGIC_KEY'] = le.fit_transform(df_encoded['MAGIC_KEY'])

#sns.set(font_scale = 2)
#plt.subplots(figsize = (10, 5))
#heat_plot = sns.heatmap(df_encoded.corr(method = 'pearson'), annot = True, cmap = 'RdYlGn', annot_kws={'size': 10})

#plt.yticks(fontsize = 10)
#plt.xticks(fontsize = 10)

#plt.show()

df_customer.info()

october_2018_first_15 = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].apply(
    lambda x: 1 if ((x.dt.month == 10) & (x.dt.day <= 15) & (x.dt.year == 2018)).any() else 0
).reset_index(name='10_2018_FIRST_15')

# Display the result DataFrame
october_2018_first_15

df_customer = pd.merge(df_customer, october_2018_first_15, on='MAGIC_KEY')
df_customer

november_2018_first_15 = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].apply(
    lambda x: 1 if ((x.dt.month == 11) & (x.dt.day <= 15) & (x.dt.year == 2018)).any() else 0
).reset_index(name='11_2018_FIRST_15')

# Display the result DataFrame
november_2018_first_15

df_customer = pd.merge(df_customer, november_2018_first_15, on='MAGIC_KEY')
df_customer

december_2018_first_15 = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].apply(
    lambda x: 1 if ((x.dt.month == 12) & (x.dt.day <= 15) & (x.dt.year == 2018)).any() else 0
).reset_index(name='12_2018_FIRST_15')

# Display the result DataFrame
december_2018_first_15

df_customer = pd.merge(df_customer, december_2018_first_15, on='MAGIC_KEY')
df_customer

january_2019_first_15 = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].apply(
    lambda x: 1 if ((x.dt.month == 1) & (x.dt.day <= 15) & (x.dt.year == 2019)).any() else 0
).reset_index(name='1_2019_FIRST_15')

# Display the result DataFrame
january_2019_first_15

df_customer = pd.merge(df_customer, january_2019_first_15, on='MAGIC_KEY')
df_customer

february_2019_first_15 = df.groupby('MAGIC_KEY')['PURCHASE_DATE'].apply(
    lambda x: 1 if ((x.dt.month == 2) & (x.dt.day <= 15) & (x.dt.year == 2019)).any() else 0
).reset_index(name='2_2019_FIRST_15')

# Display the result DataFrame
february_2019_first_15

df_customer = pd.merge(df_customer, february_2019_first_15, on='MAGIC_KEY')
df_customer

df_customer.to_csv("df_customer_all.csv", index=False)  # Save test data
